third_party/blink/web_tests/external/wpt/webxr/xr_viewport_scale.https-expected.txt

This is a testharness.js-based test.
PASS requestViewportScale valid viewport for inline session - webgl
PASS requestViewportScale valid viewport for inline session - webgl2
PASS requestViewportScale valid viewport w/ null scale for inline session - webgl
PASS requestViewportScale valid viewport w/ null scale for inline session - webgl2
PASS requestViewportScale valid viewport w/ undefined scale for inline session - webgl
PASS requestViewportScale valid viewport w/ undefined scale for inline session - webgl2
PASS requestViewportScale valid viewport w/ very small scale for inline session - webgl
PASS requestViewportScale valid viewport w/ very small scale for inline session - webgl2
PRECONDITION_FAILED requestViewportScale applied next frame for inline session - webgl requestViewportScale has no effect
PRECONDITION_FAILED requestViewportScale applied next frame for inline session - webgl2 requestViewportScale has no effect
PRECONDITION_FAILED requestViewportScale same frame for inline session - webgl requestViewportScale has no effect
PRECONDITION_FAILED requestViewportScale same frame for inline session - webgl2 requestViewportScale has no effect
PRECONDITION_FAILED recommendedViewportScale for inline session - webgl recommendedViewportScale not provided
PRECONDITION_FAILED recommendedViewportScale for inline session - webgl2 recommendedViewportScale not provided
PASS requestViewportScale valid viewport for immersive-vr session - webgl
PASS requestViewportScale valid viewport for immersive-vr session - webgl2
PASS requestViewportScale valid viewport w/ null scale for immersive-vr session - webgl
PASS requestViewportScale valid viewport w/ null scale for immersive-vr session - webgl2
PASS requestViewportScale valid viewport w/ undefined scale for immersive-vr session - webgl
PASS requestViewportScale valid viewport w/ undefined scale for immersive-vr session - webgl2
PASS requestViewportScale valid viewport w/ very small scale for immersive-vr session - webgl
PASS requestViewportScale valid viewport w/ very small scale for immersive-vr session - webgl2
PASS requestViewportScale applied next frame for immersive-vr session - webgl
PASS requestViewportScale applied next frame for immersive-vr session - webgl2
PASS requestViewportScale same frame for immersive-vr session - webgl
PASS requestViewportScale same frame for immersive-vr session - webgl2
PRECONDITION_FAILED recommendedViewportScale for immersive-vr session - webgl recommendedViewportScale not provided
PRECONDITION_FAILED recommendedViewportScale for immersive-vr session - webgl2 recommendedViewportScale not provided
Harness: the test ran to completion.
